speech_to_text_api/requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
fastapi>=0.95.0
python-multipart>=0.0.6
uvicorn>=0.21.1
pydantic>=2.0.0
openai-whisper>=20230314
python-dotenv>=1.0.0
numpy>=1.22.0
ffmpeg-python>=0.2.0